+++ packages/instrumentation-oracledb/src/instrumentation.ts
@@ -1,6 +1,6 @@
 /*
  * Copyright The OpenTelemetry Authors
- * Copyright (c) 2025, Oracle and/or its affiliates.
+ * Copyright (c) 2025, 2026, Oracle and/or its affiliates.
  * SPDX-License-Identifier: Apache-2.0
  */
 
@@ -9,7 +9,8 @@ import {
   InstrumentationNodeModuleDefinition,
 } from '@opentelemetry/instrumentation';
 import type * as oracleDBTypes from 'oracledb';
-import { OracleInstrumentationConfig } from './types';
+import type { OracleInstrumentationConfig } from './types';
+import { setMetricInstruments } from './metricUtils';
 import { getOracleTelemetryTraceHandlerClass } from './OracleTelemetryTraceHandler';
 /** @knipignore */
 import { PACKAGE_NAME, PACKAGE_VERSION } from './version';
@@ -21,10 +22,14 @@ export class OracleInstrumentation extends InstrumentationBase {
     super(PACKAGE_NAME, PACKAGE_VERSION, config);
   }
 
+  override _updateMetricInstruments(): void {
+    setMetricInstruments(this.meter);
+  }
+
   protected init() {
     const moduleOracleDB = new InstrumentationNodeModuleDefinition(
       'oracledb',
-      ['>= 6.7 < 7'],
+      ['>= 6.7 < 8'],
       (moduleExports: typeof oracleDBTypes) => {
         if (!moduleExports) {
           return;
@@ -59,8 +64,6 @@ export class OracleInstrumentation extends InstrumentationBase {
 
   override setConfig(config: OracleInstrumentationConfig = {}) {
     super.setConfig(config);
-
-    // update the config in OracleTelemetryTraceHandler obj.
     this._tmHandler?.setInstrumentConfig(this._config);
   }
 }