+++ packages/instrumentation-tedious/test/instrumentation.test.ts
@@ -11,24 +11,13 @@ import {
   type Attributes,
 } from '@opentelemetry/api';
 import { AsyncLocalStorageContextManager } from '@opentelemetry/context-async-hooks';
-import {
-  DB_SYSTEM_VALUE_MSSQL,
-  ATTR_DB_NAME,
-  ATTR_DB_SQL_TABLE,
-  ATTR_DB_STATEMENT,
-  ATTR_DB_SYSTEM,
-  ATTR_DB_USER,
-  ATTR_NET_PEER_NAME,
-  ATTR_NET_PEER_PORT,
-} from '../src/semconv';
-import { SemconvStability } from '@opentelemetry/instrumentation';
 import * as testUtils from '@opentelemetry/contrib-test-utils';
 import {
-  BasicTracerProvider,
+  TracerProvider,
   InMemorySpanExporter,
   ReadableSpan,
   SimpleSpanProcessor,
-} from '@opentelemetry/sdk-trace-base';
+} from '@opentelemetry/sdk-trace';
 import * as assert from 'assert';
 import { TediousInstrumentation } from '../src';
 import makeApi from './api';
@@ -37,17 +26,16 @@ import * as semver from 'semver';
 import {
   ATTR_DB_COLLECTION_NAME,
   ATTR_DB_NAMESPACE,
+  ATTR_DB_OPERATION_NAME,
   ATTR_DB_QUERY_TEXT,
+  ATTR_DB_RESPONSE_STATUS_CODE,
+  ATTR_DB_STORED_PROCEDURE_NAME,
   ATTR_DB_SYSTEM_NAME,
+  ATTR_ERROR_TYPE,
   ATTR_SERVER_ADDRESS,
   ATTR_SERVER_PORT,
   DB_SYSTEM_NAME_VALUE_MICROSOFT_SQL_SERVER,
 } from '@opentelemetry/semantic-conventions';
-
-// By default tests run with both old and stable semconv. Some test cases
-// specifically test the various values of OTEL_SEMCONV_STABILITY_OPT_IN.
-process.env.OTEL_SEMCONV_STABILITY_OPT_IN = 'http/dup,database/dup';
-const DEFAULT_NET_SEMCONV_STABILITY = SemconvStability.DUPLICATE;
 
 const port = Number(process.env.MSSQL_PORT) || 1433;
 const database = process.env.MSSQL_DATABASE || 'master';
@@ -90,15 +78,17 @@ const incompatVersions =
     semver.gte(tediousVersion, '17.0.0')) ||
   // tedious@19 removed support for node <18.17.0 https://github.com/tediousjs/tedious/releases/tag/v19.0.0
   (semver.lt(processVersion, '18.17.0') &&
-    semver.gte(tediousVersion, '19.0.0'));
+    semver.gte(tediousVersion, '19.0.0')) ||
+  // tedious@20 removed support for node <22 https://github.com/tediousjs/tedious/releases/tag/v20.0.0
+  (semver.lt(processVersion, '22.0.0') && semver.gte(tediousVersion, '20.0.0'));
 
 describe('tedious', () => {
   let tedious: any;
   let contextManager: AsyncLocalStorageContextManager;
   let connection: Connection;
   const memoryExporter = new InMemorySpanExporter();
-  const provider = new BasicTracerProvider({
-    spanProcessors: [new SimpleSpanProcessor(memoryExporter)],
+  const provider = new TracerProvider({
+    spanProcessors: [new SimpleSpanProcessor({ exporter: memoryExporter })],
   });
   const shouldTest = process.env.RUN_MSSQL_TESTS;
 
@@ -155,6 +145,7 @@ describe('tedious', () => {
 
     assertSpan(spans[0], {
       name: 'execSql master',
+      operationName: 'execSql',
       sql: queryString,
       parentSpan,
     });
@@ -174,8 +165,10 @@ describe('tedious', () => {
 
     assertSpan(spans[0], {
       name: 'execSql master',
+      operationName: 'execSql',
       sql: queryString,
       error: /incorrect syntax/i,
+      errorType: /^RequestError$/,
       statementCount: 0,
     });
   });
@@ -196,6 +189,7 @@ describe('tedious', () => {
 
     assertSpan(spans[0], {
       name: 'execSql master',
+      operationName: 'execSql',
       sql: queryString,
       procCount: 1,
       statementCount: 3,
@@ -213,6 +207,7 @@ describe('tedious', () => {
 
     assertSpan(spans[0], {
       name: 'execSqlBatch master',
+      operationName: 'execSqlBatch',
       sql: queryString,
       procCount: 0,
       statementCount: 3,
@@ -229,11 +224,14 @@ describe('tedious', () => {
 
     assertSpan(spans[0], {
       name: 'execSql master',
+      operationName: 'execSql',
       sql: /create or alter procedure/i,
     });
     assertSpan(spans[1], {
-      name: `callProcedure ${tedious.storedProcedure.procedureName} master`,
+      name: `EXECUTE ${tedious.storedProcedure.procedureName}`,
+      operationName: 'EXECUTE',
       sql: tedious.storedProcedure.procedureName,
+      storedProcedureName: tedious.storedProcedure.procedureName,
     });
   });
 
@@ -249,15 +247,18 @@ describe('tedious', () => {
 
     assertSpan(spans[0], {
       name: 'execSql master',
+      operationName: 'execSql',
       sql: /create table/i,
       statementCount: 2,
     });
     assertSpan(spans[1], {
       name: 'prepare master',
+      operationName: 'prepare',
       sql: /INSERT INTO/,
     });
     assertSpan(spans[2], {
       name: 'execute master',
+      operationName: 'execute',
       sql: /INSERT INTO/,
     });
   });
@@ -280,14 +281,17 @@ describe('tedious', () => {
 
     assertSpan(spans[0], {
       name: 'execSql master',
+      operationName: 'execSql',
       sql: sql.create,
     });
     assertSpan(spans[1], {
       name: 'execSql master',
+      operationName: 'execSql',
       sql: sql.use,
     });
     assertSpan(spans[2], {
       name: 'execSql temp_otel_db',
+      operationName: 'execSql',
       sql: sql.select,
       database: 'temp_otel_db',
     });
@@ -301,84 +305,21 @@ describe('tedious', () => {
 
     assertSpan(spans[0], {
       name: 'execSql master',
+      operationName: 'execSql',
       sql: /create table/i,
       statementCount: 2,
     });
     assertSpan(spans[1], {
       name: 'execSqlBatch master',
+      operationName: 'execSqlBatch',
       sql: /insert bulk/,
       procCount: 0,
     });
     assertSpan(spans[2], {
-      name: 'execBulkLoad test_bulk master',
+      name: 'BULK INSERT test_bulk',
+      operationName: 'BULK INSERT',
       procCount: 0,
       table: 'test_bulk',
-    });
-  });
-
-  describe('various values of OTEL_SEMCONV_STABILITY_OPT_IN', () => {
-    const _origOptInEnv = process.env.OTEL_SEMCONV_STABILITY_OPT_IN;
-    after(() => {
-      process.env.OTEL_SEMCONV_STABILITY_OPT_IN = _origOptInEnv;
-      (instrumentation as any)._setSemconvStabilityFromEnv();
-    });
-
-    it('OTEL_SEMCONV_STABILITY_OPT_IN=(empty)', async () => {
-      process.env.OTEL_SEMCONV_STABILITY_OPT_IN = '';
-      (instrumentation as any)._setSemconvStabilityFromEnv();
-      memoryExporter.reset();
-
-      const queryString = "SELECT 42, 'hello world'";
-      const PARENT_NAME = 'parentSpan';
-      const parentSpan = provider.getTracer('default').startSpan(PARENT_NAME);
-      assert.deepStrictEqual(
-        await context.with(trace.setSpan(context.active(), parentSpan), () =>
-          tedious.query(connection, queryString)
-        ),
-        [42, 'hello world']
-      );
-      parentSpan.end();
-      const spans = memoryExporter.getFinishedSpans();
-      assert.strictEqual(spans.length, 2, 'Received incorrect number of spans');
-      assertSpan(
-        spans[0],
-        {
-          name: 'execSql master',
-          sql: queryString,
-          parentSpan,
-        },
-        SemconvStability.OLD
-      );
-      assert.strictEqual(spans[1].name, PARENT_NAME);
-    });
-
-    it('OTEL_SEMCONV_STABILITY_OPT_IN=http,database', async () => {
-      process.env.OTEL_SEMCONV_STABILITY_OPT_IN = 'http,database';
-      (instrumentation as any)._setSemconvStabilityFromEnv();
-      memoryExporter.reset();
-
-      const queryString = "SELECT 42, 'hello world'";
-      const PARENT_NAME = 'parentSpan';
-      const parentSpan = provider.getTracer('default').startSpan(PARENT_NAME);
-      assert.deepStrictEqual(
-        await context.with(trace.setSpan(context.active(), parentSpan), () =>
-          tedious.query(connection, queryString)
-        ),
-        [42, 'hello world']
-      );
-      parentSpan.end();
-      const spans = memoryExporter.getFinishedSpans();
-      assert.strictEqual(spans.length, 2, 'Received incorrect number of spans');
-      assertSpan(
-        spans[0],
-        {
-          name: 'execSql master',
-          sql: queryString,
-          parentSpan,
-        },
-        SemconvStability.STABLE
-      );
-      assert.strictEqual(spans[1].name, PARENT_NAME);
     });
   });
 
@@ -465,11 +406,7 @@ const assertRejects = (
     });
 };
 
-function assertSpan(
-  span: ReadableSpan,
-  expected: any,
-  semconvStability: SemconvStability = DEFAULT_NET_SEMCONV_STABILITY
-) {
+function assertSpan(span: ReadableSpan, expected: any) {
   assert.ok(span);
   assert.strictEqual(span.name, expected.name);
   assert.strictEqual(span.kind, SpanKind.CLIENT);
@@ -480,62 +417,54 @@ function assertSpan(
     'tedious.procedure_count': expected.procCount ?? 1,
     'tedious.statement_count': expected.statementCount ?? 1,
   };
-  if (semconvStability & SemconvStability.OLD) {
-    expectedAttrs[ATTR_DB_SYSTEM] = DB_SYSTEM_VALUE_MSSQL;
-    expectedAttrs[ATTR_DB_NAME] = expected.database ?? database;
-    expectedAttrs[ATTR_DB_USER] = user;
-    expectedAttrs[ATTR_NET_PEER_NAME] = host;
-    expectedAttrs[ATTR_NET_PEER_PORT] = port;
-    if (expected.table) {
-      expectedAttrs[ATTR_DB_SQL_TABLE] = expected.table;
-    }
-    // "db.statement"
-    if (expected.sql) {
-      if (expected.sql instanceof RegExp) {
-        assert.match(
-          span.attributes[ATTR_DB_STATEMENT] as string,
-          expected.sql
-        );
-      } else {
-        assert.strictEqual(
-          span.attributes[ATTR_DB_STATEMENT],
-          expected.sql,
-          ATTR_DB_STATEMENT
-        );
-      }
-    } else {
-      assert.strictEqual(actualAttrs[ATTR_DB_STATEMENT], undefined);
-    }
-    delete actualAttrs[ATTR_DB_STATEMENT];
+
+  expectedAttrs[ATTR_DB_SYSTEM_NAME] =
+    DB_SYSTEM_NAME_VALUE_MICROSOFT_SQL_SERVER;
+  expectedAttrs[ATTR_DB_NAMESPACE] = expected.database ?? database;
+  expectedAttrs[ATTR_SERVER_ADDRESS] = host;
+  expectedAttrs[ATTR_SERVER_PORT] = port;
+  if (expected.table) {
+    expectedAttrs[ATTR_DB_COLLECTION_NAME] = expected.table;
   }
-  if (semconvStability & SemconvStability.STABLE) {
-    expectedAttrs[ATTR_DB_SYSTEM_NAME] =
-      DB_SYSTEM_NAME_VALUE_MICROSOFT_SQL_SERVER;
-    expectedAttrs[ATTR_DB_NAMESPACE] = expected.database ?? database;
-    expectedAttrs[ATTR_SERVER_ADDRESS] = host;
-    expectedAttrs[ATTR_SERVER_PORT] = port;
-    if (expected.table) {
-      expectedAttrs[ATTR_DB_COLLECTION_NAME] = expected.table;
-    }
-    // "db.statement"
-    if (expected.sql) {
-      if (expected.sql instanceof RegExp) {
-        assert.match(
-          span.attributes[ATTR_DB_QUERY_TEXT] as string,
-          expected.sql
-        );
-      } else {
-        assert.strictEqual(
-          span.attributes[ATTR_DB_QUERY_TEXT],
-          expected.sql,
-          ATTR_DB_QUERY_TEXT
-        );
-      }
-    } else {
-      assert.strictEqual(actualAttrs[ATTR_DB_QUERY_TEXT], undefined);
-    }
-    delete actualAttrs[ATTR_DB_QUERY_TEXT];
+  if (expected.operationName) {
+    expectedAttrs[ATTR_DB_OPERATION_NAME] = expected.operationName;
   }
+  if (expected.storedProcedureName) {
+    expectedAttrs[ATTR_DB_STORED_PROCEDURE_NAME] = expected.storedProcedureName;
+  }
+
+  // db.query.text
+  if (expected.sql) {
+    if (expected.sql instanceof RegExp) {
+      assert.match(span.attributes[ATTR_DB_QUERY_TEXT] as string, expected.sql);
+    } else {
+      assert.strictEqual(
+        span.attributes[ATTR_DB_QUERY_TEXT],
+        expected.sql,
+        ATTR_DB_QUERY_TEXT
+      );
+    }
+  } else {
+    assert.strictEqual(actualAttrs[ATTR_DB_QUERY_TEXT], undefined);
+  }
+  delete actualAttrs[ATTR_DB_QUERY_TEXT];
+
+  if (expected.errorType) {
+    assert.match(
+      span.attributes[ATTR_ERROR_TYPE] as string,
+      expected.errorType
+    );
+  }
+  delete actualAttrs[ATTR_ERROR_TYPE];
+
+  if (expected.statusCode) {
+    assert.match(
+      span.attributes[ATTR_DB_RESPONSE_STATUS_CODE] as string,
+      expected.statusCode
+    );
+  }
+  delete actualAttrs[ATTR_DB_RESPONSE_STATUS_CODE];
+
   assert.deepEqual(actualAttrs, expectedAttrs);
 
   if (expected.parentSpan) {